scripts/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
supabase>=1.0.0
schedule>=1.2.0
python-dotenv>=0.19.0
lxml>=4.9.0 